Releasing

+++ Service/ProductService.cs
@@ -1,36 +1,46 @@
 using apicatalogo.Models;
-using ApiCatalogo.Repositories;
+using ApiCatalogo.DTOs;
+using AutoMapper;
 
 namespace ApiCatalogo.Service
 {
     public class ProductService : IProductService
     {
         private readonly IUnitOfWork _uof;
+        private readonly IMapper _mapper;
 
-        public ProductService(IUnitOfWork uof)
+        public ProductService(IUnitOfWork uof, IMapper mapper)
         {
             _uof = uof;
+            _mapper = mapper;
 
         }
 
-        public Product Create(Product product)
+        public ProductDTO Create(ProductDTO product)
         {
             _ = product ?? throw new ArgumentNullException(nameof(product));
 
-            var productResponse = _uof.ProductRepository.Create(product);
+            var productentity = _mapper.Map<Product>(product);
+
+            var productResponse = _uof.ProductRepository.Create(productentity);
             _uof.Commit();
 
-            return productResponse;
+            var response = _mapper.Map<ProductDTO>(productResponse);
+
+            return response;
         }
-        public Product GetProductById(int id)
+        public ProductDTO GetProductById(int id)
         {
             var product = _uof.ProductRepository.Get(c => c.ProductId == id);
 
             _ = product ?? throw new InvalidOperationException("Produto é null");
 
-            return product;
+            // _mapper.Map<Destino>(Origem);
+            var response = _mapper.Map<ProductDTO>(product);
+
+            return response;
         }
-        public Product? Update(Product product)
+        public ProductDTO? Update(ProductDTO product)
         {
             _ = product ?? throw new ArgumentNullException(nameof(product));
 
@@ -38,31 +48,40 @@ namespace ApiCatalogo.Service
 
             if (hasProduct is not null)
             {
-                var productModel = _uof.ProductRepository.Update(product);
+                var productEntity = _mapper.Map<Product>(product);
+                var productModel = _uof.ProductRepository.Update(productEntity);
                 _uof.Commit();
 
-                return productModel;
+                var response = _mapper.Map<ProductDTO>(productModel);
+
+                return response;
             }
 
             return null;
 
         }
-        public Product Delete(int id)
+        public ProductDTO Delete(int id)
         {
             var product = _uof.ProductRepository.Get(c => c.ProductId == id);
-            _ = product ?? null;
+            if (product is null) return null;
 
-            return _uof.ProductRepository.Delete(product!);
+            var productEntity = _uof.ProductRepository.Delete(product!);
+            var response = _mapper.Map<ProductDTO>(productEntity);
+
+            return response;
         }
-        public IEnumerable<Product> GetProducts()
+        public IEnumerable<ProductDTO> GetProducts()
         {
-            return _uof.ProductRepository.GetAll();
+            IEnumerable<Product> productEntity = _uof.ProductRepository.GetAll();
+            var response = _mapper.Map<IEnumerable<ProductDTO>>(productEntity);
+            return response;
         }
-        public IEnumerable<Product> GetProductByCategory(int id)
+        public IEnumerable<ProductDTO> GetProductByCategory(int id)
         {
             var products = _uof.ProductRepository.GetProductsByCategory(id);
-            _ = products ?? null;
-            return products!;
+            var response = _mapper.Map<IEnumerable<ProductDTO>>(products);
+
+            return response;
         }
